Improve promptflow-tracing tests
